fix: Remove unsafety by moving Executable to a heap vector
