+++ nova_vm/src/engine/bytecode.rs
@@ -2,16 +2,17 @@
 // License, v. 2.0. If a copy of the MPL was not distributed with this
 // file, You can obtain one at https://mozilla.org/MPL/2.0/.
 
+mod bytecode_compiler;
 mod executable;
-mod heap_allocated_bytecode;
 mod instructions;
 pub(super) mod iterator;
 mod vm;
 
-pub(crate) use executable::{
-    is_reference, CompileContext, CompileEvaluation, Executable, FunctionExpression, IndexType,
-    NamedEvaluationParameter, SendableRef,
+pub(crate) use bytecode_compiler::{
+    is_reference, CompileContext, CompileEvaluation, NamedEvaluationParameter,
 };
-pub(crate) use heap_allocated_bytecode::HeapAllocatedBytecode;
+pub(crate) use executable::{
+    Executable, ExecutableHeapData, FunctionExpression, IndexType, SendableRef,
+};
 pub(crate) use instructions::{Instruction, InstructionIter};
 pub(crate) use vm::{instanceof_operator, ExecutionResult, SuspendedVm, Vm};
