+++ nova_vm/src/engine/bytecode/bytecode_compiler/class_definition_evaluation.rs
