+++ nova_vm/src/heap.rs
@@ -43,18 +43,6 @@ use crate::ecmascript::builtins::{
     weak_set::data::WeakSetHeapData,
 };
 use crate::ecmascript::{
-    builtins::ArrayHeapData,
-    execution::{Environments, Realm, RealmIdentifier},
-    scripts_and_modules::{
-        module::ModuleIdentifier,
-        script::{Script, ScriptIdentifier},
-    },
-    types::{
-        BigIntHeapData, BoundFunctionHeapData, BuiltinFunctionHeapData, ECMAScriptFunctionHeapData,
-        NumberHeapData, Object, ObjectHeapData, String, StringHeapData, SymbolHeapData, Value,
-    },
-};
-use crate::ecmascript::{
     builtins::{
         control_abstraction_objects::{
             async_function_objects::await_reaction::AwaitReaction,
@@ -63,14 +51,6 @@ use crate::ecmascript::{
                 promise_reaction_records::PromiseReactionRecord,
                 promise_resolving_functions::PromiseResolvingFunctionHeapData,
             },
-        },
-        embedder_object::data::EmbedderObjectHeapData,
-        error::ErrorHeapData,
-        finalization_registry::data::FinalizationRegistryHeapData,
-        indexed_collections::array_objects::array_iterator_objects::array_iterator::ArrayIteratorHeapData,
-        keyed_collections::{
-            map_objects::map_iterator_objects::map_iterator::MapIteratorHeapData,
-            set_objects::set_iterator_objects::set_iterator::SetIteratorHeapData,
         },
         map::data::MapHeapData,
         module::data::ModuleHeapData,
@@ -85,6 +65,32 @@ use crate::ecmascript::{
         bigint::HeapBigInt, BuiltinConstructorHeapData, HeapNumber, HeapString, OrdinaryObject,
         BUILTIN_STRINGS_LIST,
     },
+};
+use crate::{
+    ecmascript::{
+        builtins::{
+            embedder_object::data::EmbedderObjectHeapData,
+            error::ErrorHeapData,
+            finalization_registry::data::FinalizationRegistryHeapData,
+            indexed_collections::array_objects::array_iterator_objects::array_iterator::ArrayIteratorHeapData,
+            keyed_collections::{
+                map_objects::map_iterator_objects::map_iterator::MapIteratorHeapData,
+                set_objects::set_iterator_objects::set_iterator::SetIteratorHeapData,
+            },
+            ArrayHeapData,
+        },
+        execution::{Environments, Realm, RealmIdentifier},
+        scripts_and_modules::{
+            module::ModuleIdentifier,
+            script::{Script, ScriptIdentifier},
+        },
+        types::{
+            BigIntHeapData, BoundFunctionHeapData, BuiltinFunctionHeapData,
+            ECMAScriptFunctionHeapData, NumberHeapData, Object, ObjectHeapData, String,
+            StringHeapData, SymbolHeapData, Value,
+        },
+    },
+    engine::ExecutableHeapData,
 };
 pub(crate) use heap_bits::{CompactionLists, HeapMarkAndSweep, WorkQueues};
 
@@ -111,6 +117,8 @@ pub struct Heap {
     pub embedder_objects: Vec<Option<EmbedderObjectHeapData>>,
     pub environments: Environments,
     pub errors: Vec<Option<ErrorHeapData>>,
+    /// Stores compiled bytecodes
+    pub(crate) executables: Vec<ExecutableHeapData>,
     pub finalization_registrys: Vec<Option<FinalizationRegistryHeapData>>,
     pub generators: Vec<Option<GeneratorHeapData>>,
     pub globals: Vec<Option<Value>>,
@@ -206,6 +214,7 @@ impl Heap {
             embedder_objects: Vec::with_capacity(0),
             environments: Default::default(),
             errors: Vec::with_capacity(1024),
+            executables: Vec::with_capacity(1024),
             source_codes: Vec::with_capacity(0),
             finalization_registrys: Vec::with_capacity(0),
             generators: Vec::with_capacity(1024),
